Create Empire State Bet
The Empire state building by itself has 1860 steps of which 1576 steps are open to public for climbing only during the Empire State Building run. We consider only 60 steps for our scope and start from ground level. The idea is to roll a dice and based on some rules take the number of steps either forward or backward. At the end of the 100th Dice roll, we record the no. of steps we have covered. We then simulate this random walk game multiple number of times to finally check how many times we achieve Step 60 and the odds.